sort

+++ src/util/CheckInterface.kt
@@ -1,10 +1,8 @@
 package util
 
-import data.OperatorData
-
 interface CheckInterface {
     fun checkName(name : String): Boolean
     fun checkPhone(phone : String): Boolean
-    fun checkBalance(balance : Double): Boolean
+    fun checkBalance(balance: Int): Boolean
 //    fun checkAll(date : OperatorData): Boolean
 }
